requirements.txt
Django>=1.6
Fabric==1.8.0
South==0.8.2
argparse==1.2.1
boto==2.15.0
chardet==2.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==0.10.2
django-tastypie==0.10.0
ecdsa==0.10
gunicorn==18.0
mimeparse==0.1.3
paramiko==1.12.0
psycopg2==2.5.1
pycrypto==2.6.1
python-dateutil==2.2
python-memcached==1.53
six==1.4.1
sqlparse==0.1.10
static==0.4
wsgiref==0.1.2
ipython
TwitterAPI
